// backend/src/modules/shipment/events/shipment.event.types.ts
export enum ShipmentEventType {
  SHIPMENT_CREATED = 'shipment.created',
  SHIPMENT_UPDATED = 'shipment.updated',
  ORDER_SHIPPED = 'shipment.order_shipped',
  TRACKING_UPDATED = 'shipment.tracking_updated',
  ORDER_DELIVERED = 'shipment.order_delivered',
  SHIPMENT_CANCELLED = 'shipment.cancelled',
}